+++ frontend/src/UI/components/Controllers/DatePicker.tsx
@@ -6,9 +6,7 @@ import {
   FieldValues,
   UseControllerProps,
 } from "react-hook-form";
-import ErrorMessage from "./ErrorMessage";
-import HelperText from "./HelperText";
-import Label from "./Label";
+import { ErrorMessage, HelperText, Label } from "../Text";
 
 type DatePickerProps<T extends FieldValues> = {
   error?: FieldError | undefined;
@@ -18,7 +16,7 @@ type DatePickerProps<T extends FieldValues> = {
   required?: boolean;
 } & UseControllerProps<T>;
 
-function ControlledDatePicker<T extends FieldValues>({
+export function ControlledDatePicker<T extends FieldValues>({
   control,
   error,
   helperText,
@@ -44,5 +42,3 @@ function ControlledDatePicker<T extends FieldValues>({
     </FormControl>
   );
 }
-
-export default ControlledDatePicker;
