+++ frontend/src/UI/components/Controllers/index.tsx
@@ -1,0 +1,6 @@
+export * from "./DatePicker";
+export * from "./NumberInput";
+export * from "./Select";
+export * from "./StarRating";
+export * from "./TextArea";
+export * from "./TextInput";
