--- frontend/src/UI/components/Navigation/types.ts
@@ -1,6 +1,0 @@
-export type NavItem = {
-  label: string;
-  subLabel?: string;
-  children?: Array<NavItem>;
-  href?: string;
-};
