+++ frontend/src/UI/components/Text/HelperText.tsx
@@ -4,12 +4,10 @@ type Props = {
   text: string;
 };
 
-function HelperText({ text }: Props) {
+export function HelperText({ text }: Props) {
   return (
     <FormHelperText fontSize="xs" color="gray.400" mb={2}>
       {text}
     </FormHelperText>
   );
 }
-
-export default HelperText;
