+++ frontend/src/UI/components/Text/Label.tsx
@@ -5,12 +5,10 @@ type Props = {
   isRequired: boolean;
 };
 
-function Label({ text, isRequired }: Props) {
+export function Label({ text, isRequired }: Props) {
   return (
     <FormLabel style={{ color: "#a3a3a3", fontStyle: "italic" }}>
       {isRequired ? `${text}*` : text}
     </FormLabel>
   );
 }
-
-export default Label;
