+++ frontend/src/UI/components/Text/index.ts
@@ -1,0 +1,5 @@
+export * from "./ErrorMessage";
+export * from "./HelperText";
+export * from "./Label";
+export * from "./PageTitle";
+export * from "./Warning";
