+++ frontend/src/UI/pages/Books/BooksTable.tsx
@@ -1,5 +1,5 @@
 import { Box, Input, Spinner } from "@chakra-ui/react";
-import { genericSearch } from "../../../business/functions/genericSearch";
+import { genericSearch } from "../../../business/functions";
 import {
   stateSectionsWithTable,
   useActions,
@@ -7,9 +7,7 @@ import {
 } from "../../../business/overmind";
 import Pagination from "../../components/Table/Pagination";
 import TableOverview from "../../components/Table/TableOverView";
-import { useGetBooks } from "../../hooks/useGetBooks";
-import { useGetGenres } from "../../hooks/useGetGenres";
-import { useGetLanguages } from "../../hooks/useGetLanguages";
+import { useGetBooks, useGetGenres, useGetLanguages } from "../../hooks";
 import Filter from "./Filter";
 
 function BooksTable() {
