--- frontend/src/UI/pages/Books/helpers.ts
@@ -1,16 +1,0 @@
-import { z } from "zod";
-
-export type FormFields = {
-  genre: string;
-  language: string;
-};
-
-export const validationSchema = z.object({
-  genre: z.string(),
-  language: z.string(),
-});
-
-export const defaultValues = {
-  genre: "",
-  language: "",
-};
