+++ frontend/src/UI/pages/Reviews/AddReviewPage.tsx
@@ -3,19 +3,20 @@ import { DevTool } from "@hookform/devtools";
 import { zodResolver } from "@hookform/resolvers/zod";
 import { useEffect, useState } from "react";
 import { SubmitHandler, useForm } from "react-hook-form";
-import { calculateDays } from "../../../business/functions/date";
+import { calculateDays } from "../../../business/functions";
 import { useAppState } from "../../../business/overmind";
-import ControlledDatePicker from "../../components/Controllers/DatePicker";
-import ControlledNumberInput from "../../components/Controllers/NumberInput";
-import ControlledSelect from "../../components/Controllers/Select";
+import {
+  ControlledDatePicker,
+  ControlledNumberInput,
+  ControlledSelect,
+  ControlledStarRating,
+  ControlledTextArea,
+  ControlledTextInput,
+} from "../../components/Controllers";
 // import { useActions } from "../../../business/overmind";
-import ControlledStarRating from "../../components/Controllers/StarRating";
-import ControlledTextArea from "../../components/Controllers/TextArea";
-import ControlledTextInput from "../../components/Controllers/TextInput";
 import NavigationBar from "../../components/Navigation/NavigationBar";
-import PageTitle from "../../components/Text/PageTitle";
-import { useGetAuthors } from "../../hooks/useGetAuthors";
-import { useGetBooks } from "../../hooks/useGetBooks";
+import { PageTitle } from "../../components/Text";
+import { useGetAuthors, useGetBooks } from "../../hooks";
 import { defaultValues, FormFields, validationSchema } from "./helpers";
 
 function AddReviewPage() {
