+++ frontend/src/business/functions/index.ts
@@ -1,0 +1,5 @@
+export * from "./compare";
+export * from "./date";
+export * from "./genericSearch";
+export * from "./getColorIndex";
+export * from "./groupBy";
